refactor: correction in controller import
