+++ src/main/java/br/com/matheusgusmao/incometax/web/dto/dependent/DependentResponse.java
@@ -2,6 +2,8 @@ package br.com.matheusgusmao.incometax.web.dto.dependent;
 
 import br.com.matheusgusmao.incometax.domain.model.dependent.Dependent;
 
+import java.time.LocalDate;
+
 public record DependentResponse(Long id, String name, String cpf, LocalDate birthDate) {
 
     public static DependentResponse from(Dependent dependent) {
